Add initial support for V2 streaming API.
This adds 2 new variables to `test` and `collate` functions: 'stream' (bool) and 'timeout' (request timeout duration)
Intended usage so far is to use 16.5kb compressed bodies to simulate agent behaviour.

All functionally depending on accepted events is commented out, as we need to read the response from the apm-server to handle that.
In the meanwhile, tests that test too much are skipped - they could be removed later.

Co-authored-by: Silvia Mitter <silvia.mitter@elastic.co>